requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
requests==2.26.0
furl==2.1.3
waitress==2.0.0
Paste==3.5.2
requests-oauthlib==2.0.0
pytz==2025.1
cachetools==5.5.2
beautifulsoup4==4.11.0
pyyaml!=6.0.0,!=5.4.0,!=5.4.1
Werkzeug>=2.2,<3.0
